refactor: move error rendering  to exception handler
