+++ app/Exceptions/CustomException.php
@@ -1,5 +1,5 @@
 <?php
 
-namespace App\Components\Errors;
+namespace App\Exceptions;
 
 class CustomException extends \Exception {}
